+++ NEWS.md
@@ -1,3 +1,5 @@
+# occumb (development version)
+
 # occumb v0.2.1 (2021/11/21)
 * A few bug fixes.
 * Add validations for the inputs of `occumb()`.