+++ NEWS.md
@@ -3,6 +3,7 @@
 * Add `get_data()` function.
 * Add `get_post_sample()` function.
 * Add `gof()` function.
+* Add `data` field in `occumbFit` class.
 
 # occumb v0.2.1 (2021/11/21)
 * A few bug fixes.